+++ test/e2e/config/rabbitmq/rabbitmq.go
@@ -1,5 +1,3 @@
-// +build e2e
-
 /*
 Copyright 2020 The Knative Authors
