+++ test/e2e/smoke_test_broker_trigger.go
@@ -1,5 +1,3 @@
-// +build e2e
-
 /*
 Copyright 2020 The Knative Authors
 
@@ -19,9 +17,21 @@ limitations under the License.
 package e2e
 
 import (
+	"context"
+	"strings"
+	"testing"
+	"time"
+
 	"knative.dev/eventing-rabbitmq/test/e2e/config/rabbitmq"
 	"knative.dev/eventing-rabbitmq/test/e2e/config/smoke/brokertrigger"
+	"knative.dev/reconciler-test/pkg/environment"
 	"knative.dev/reconciler-test/pkg/feature"
+	"knative.dev/reconciler-test/pkg/k8s"
+)
+
+const (
+	interval = 1 * time.Second
+	timeout  = 5 * time.Minute
 )
 
 // SmokeTestBrokerImpl makes sure an RabbitMQ Broker goes ready.
@@ -32,4 +42,18 @@ func SmokeTestBrokerTrigger() *feature.Feature {
 	f.Setup("install a broker", brokertrigger.Install())
 	f.Alpha("RabbitMQ broker").Must("goes ready", AllGoReady)
 	return f
+}
+
+func AllGoReady(ctx context.Context, t *testing.T) {
+	env := environment.FromContext(ctx)
+	for _, ref := range env.References() {
+		if !strings.Contains(ref.APIVersion, "knative.dev") {
+			// Let's not care so much about checking the status of non-Knative
+			// resources.
+			continue
+		}
+		if err := k8s.WaitForReadyOrDone(ctx, ref, interval, timeout); err != nil {
+			t.Fatal("failed to wait for ready or done, ", err)
+		}
+	}
 }
